+++ packages/editor/actions/user.js
@@ -211,6 +211,6 @@ export function requestResetPassword(email) {
   return async () => axios.post('/api/email/reset/request', { email });
 }
 
-export function registerEmail(email, password) {
-  return async () => axios.post('/api/email', { email, password });
+export function registerEmail(email, password, organization) {
+  return async () => axios.post('/api/email', { email, password, organization });
 }